// src/membership/dto/create-membership.dto.ts
export class CreateMembershipDto {
    membershipId?: string;
    membershipImageUrl: string;
    membershipDescription?: string;
    membershipContractSymbol?: string;
    membershipName?: string;
    membershipContractAddress?: string;
    nftBlockId?: string;
    nftBlockMintedBlockAddress?: string;
    userId?: string;
    brandId?: string;
  }